Practicing module pattern. Finished basic logic
Trying to organize my code into modules and
function factories that keep my code organized
and less/not at all reliant on global
variables. I am constructing the game to work
with basic console commands, so havent done
anything to HTML and CSS yet. Need to finish
the game checking logic to check for wins, ties,
and if a space is valid for marking.