Finished the win condition checking
I didn't know how to approach this at first,
but I decided to just make a temporary array
that would store all of the directions to
use the .every() array method to check if
they were all the same. If they were, then it
would declare the current player the winner.