Finished DOM IIFE Module for DOM manipulation
Coded the DOM module to make buttons based on
the gameboard array, and playturns anytime the
button is pressed with the button handler, and
it deletes the current game board and updates it
to the new one with the player's token. Also
added a div to hold the button for styling
the lines in between. Need to add more
styling and additional functionalities
as stated in the odin project assignment.